fix(config): fix module resolution after rename
