--- src/Config/index.ts
@@ -1,3 +1,0 @@
-export type * from '@/config/koala-config';
-export * from '@/config/default-config';
-export * from '@/config/config-loader';
